requirements.txt
absl-py==1.4.0
cachetools==5.3.0
certifi==2023.7.22
charset-normalizer==3.1.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.53.0
idna==3.4
importlib-metadata==6.0.0
Markdown==3.4.1
MarkupSafe==2.1.2
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
Pillow==10.0.1
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6

## pytorch 1.13.1
torch==1.13.1
## torchvision==0.11.2+cu102
torchvision==0.14.1
## 新增
torchaudio==0.13.1

# CUDA 10.2
# pip install torch==1.10.1+cu102 torchvision==0.11.2+cu102 torchaudio==0.10.1 -f https://download.pytorch.org/whl/cu102/torch_stable.html
## cuda 10.2 pytorch 1.10.1
#torch==1.10.1+cu102
#torchvision==0.11.2+cu102
#torchaudio==0.10.1

typing_extensions==4.5.0
urllib3==1.26.18
Werkzeug==2.3.8
zipp==3.15.0
tqdm
opencv-python
basemap
scikit-learn
